Re-organize plugin per current best practices
