+++ commands/connect/pause.js
@@ -1,14 +1,14 @@
 'use strict'
-const api = require('./shared.js')
-const regions = require('./regions.js')
+const api = require('../../lib/connect/api.js')
+const regions = require('../../lib/connect/regions.js')
 const cli = require('heroku-cli-util')
 const co = require('co')
 
 module.exports = {
   topic: 'connect',
-  command: 'resume',
-  description: 'Resume a connection',
-  help: 'Resumes a paused connection',
+  command: 'pause',
+  description: 'Pause a connection',
+  help: 'Pauses an active connection',
   flags: [
     {name: 'resource', description: 'specific connection resource name', hasValue: true},
     regions.flag
@@ -17,9 +17,9 @@ module.exports = {
   needsAuth: true,
   run: cli.command(co.wrap(function * (context, heroku) {
     context.region = regions.determineRegion(context)
-    cli.action('resuming connection', co(function * () {
+    yield cli.action('pausing connection', co(function * () {
       let connection = yield api.withConnection(context, heroku)
-      let url = '/api/v3/connections/' + connection.id + '/actions/resume'
+      let url = '/api/v3/connections/' + connection.id + '/actions/pause'
       yield api.request(context, 'POST', url)
     }))
   }))
