+++ commands/connect/state.js
@@ -1,6 +1,6 @@
 'use strict'
-const api = require('./shared.js')
-const regions = require('./regions.js')
+const api = require('../../lib/connect/api.js')
+const regions = require('../../lib/connect/regions.js')
 const cli = require('heroku-cli-util')
 const co = require('co')
 
@@ -9,7 +9,6 @@ function * run (context, heroku) {
   let connections = yield api.withUserConnections(context, context.app, context.flags, heroku)
 
   connections.forEach(function (connection) {
-    console.log('Foo')
     cli.log(connection.state)
   })
 }
