+++ lib/connect/regions.js
